Renamed SVD->DenseSVD, deleted LanczosSVD, expanded code in dispatch
