+++ tests/linalg/svd/test_svd.py
@@ -1,6 +1,7 @@
 from cola.backends import all_backends
 from cola.fns import lazify
 from cola.linalg.algorithm_base import Auto
+from cola.linalg.decompositions.decompositions import Lanczos
 from cola.linalg.svd.svd import svd
 from cola.utils.test_utils import generate_pd_from_diag, generate_spectrum, get_xnp, parametrize, relative_error
 
@@ -15,7 +16,7 @@ def test_svd(backend):
     A = lazify(xnp.array(generate_pd_from_diag(diag, dtype=diag.dtype, seed=21), dtype=dtype, device=None))
     soln_vals = xnp.sort(xnp.array(diag, dtype=dtype, device=None))
 
-    U, Sigma, V = svd(A, Auto(tol=1e-6, method="dense"))
+    U, Sigma, V = svd(A, Auto(tol=1e-6))
 
     rel_error = relative_error(soln_vals, Sigma.diag)
     print(f"Rel error: {rel_error:2.5e}")
@@ -25,7 +26,7 @@ def test_svd(backend):
     print(f"Rel error: {rel_error:2.5e}")
     assert rel_error < _tol * 5
 
-    U, Sigma, V = svd(A, Auto(tol=1e-6, method="iterative"))
+    U, Sigma, V = svd(A, Lanczos(tol=1e-6))
 
     rel_error = relative_error(soln_vals, Sigma.diag)
     print(f"Rel error: {rel_error:2.5e}")
